Add Template test for `index` template
Add a Template test for the `index` template for the Twatter application,
which includes its layout and such in order to test the rendered
template in isolation. However, since the `erb` method is part of the
Base Sinatra application it needs to exist or be running to rely on its
`erb` method, instead it has been recreated more simply using the Tilt
library used within Sinatra to help test it when isolated.