Fix Navigation Template test
Fix the Navigation Template test regular expression, since the Username
is displayed, rather than just the User's actual name, and because it is
now wrapped in a link as well to quickly get the User's profile page.

Most of the work in this commit is to get the Regex within 80 characters
so that Rubocop would not complain.

== Notes:
- Contains attempt to fix Blocked Users Controller test