Fix all broken tests after site behaviour change
Update & Fix all the broken tests that were broken to the change in the
website behaviour to only display followed User's Twats on the main page
when logged in, rather than all public Twats. Instead the majority of
tests first go to the User's page first and correctly makes its
assertions.

== Notes:
- Remove any duplicate Users that share unique attributes throughout
  tests

- Remove Mock User classes when not used properly when creating or
  signing in Users, it was meant for signing up Users from the UI