Add Messages model to application
Add a Message model to the application, in order fulfil the direct
messaging feature of the application, based off the existing Twitter
feature set. A Message will only an User to send a message directly to
another User, without anyone else being able to see that content, making
it private to those involved, whilst allowing a two-person dialog within
the application.

== Notes:
- This is the first attempt of using MiniTest Specs within this project